Implement edit menu
This is intended for manual TOTP profile entry (no QR codes
involved) or for cleaning up data from a problematic QR code URI.

The edit menu also allows setting a "title" field which I plan to
use for the hardware token profile title (to be shown on LCD next
to the OTP code).

Input validation is very minimal. This is meant to let you set the
values exactly how you want them. So, the burden is on the person
doing the editing to avoid setting weird values. The menu does
give some reminders about which values are suitable.