Reformat encoder code for better folding
This groups the stages of my encoder into blocks that can be folded
closed in a text editor that supports code folding. The blocks also
help to enforce limited scope for all the named const values that
I'm declaring along the way.

Perhaps it might be better to split some of this into functions,
but, for now, I like having the whole algorithm laid out in a
straight line. Helps to follow what's going on without having to
scroll and jump around.